Some questions for @petrelharp.
This shows the depth of my ignorance. :)

I'd like to better understand how you are approaching likelihood
calculation that has some per-site component but also a nontrivial
motif component. For example, in the CpG model, mutations can arise
because of the "usual" per-nucleotide mutation, or arise because of the
actual CpG motif mutation as you describe. Do we think of this as being
a model with m=1 and l=1? If that's the case then we have many
overlapping motifs, and so then do we only calculate the likelihood for
part of the data via nonoverlapping windows? Or do we choose a clever
parametrization like suggested in 3.2 that has mutually excluding
patterns?