Adding aggregates to summarize the whole sim.
I add a little `summarize-across.R` script to make a summary of the
results of all of the sims, e.g. `_seed/summary-across.tsv`. Right now
it calculates covariance, which is stupid, but we can think about what
is better.

The script gets run on the collection of `complete.Rdata` files that are
generated in the code via the aggregation functionality of nestly.

Closes #27.